+++ src/test/java/im/mak/paddle/blockchain_updates/transactions_checkers/invoke_transactions_checkers/InvokeStateUpdateAssertions.java
@@ -1,4 +1,4 @@
-package im.mak.paddle.blockchain_updates.subscribe_tests.subscribe_invoke_tx_tests.invoke_transactions_checkers;
+package im.mak.paddle.blockchain_updates.transactions_checkers.invoke_transactions_checkers;
 
 import java.util.Map;
